Note change needed to push to CF using Java 11
Another developer and I ran into an issue where `cf push` failed with a message referring to jvm version mismatches. After some investigation, we discovered that we just needed to uncomment a single line in `manifest.yml`. It seems to me that it would be valuable to have that guidance up front, rather than forcing users to discover it. I suspect that many people are likely to just move on to another sample app if `cf push` fails unexpectedly like that.